refactor: split server.go into modular handler files
